+++ src/main/kotlin/icu/windea/pls/config/util/CwtConfigManipulator.kt
@@ -146,7 +146,6 @@ object CwtConfigManipulator {
             else -> return null
         }
     }
-
     //endregion
     
     //region Merge Methods